Refactor tenant api's into sub-app
